+++ src/main/services/team/provisioning/TeamProvisioningOpenCodeMemberIdentityBoundaryFactory.ts
@@ -6,7 +6,7 @@ import {
 import type {
   OpenCodeMemberDirectory,
   OpenCodeMemberIdentityResolution,
-} from '../opencode/delivery/OpenCodeMemberMessageDeliveryService';
+} from '../opencode/delivery/OpenCodeMemberMessageDeliveryPorts';
 import type { TeamProviderId } from '@shared/types';
 
 export interface TeamProvisioningOpenCodeMemberIdentityBoundaryDeps {
