+++ src/main/services/team/provisioning/TeamProvisioningOpenCodeMemberInboxRelayResults.ts
@@ -1,4 +1,4 @@
-import type { OpenCodeMemberInboxDelivery } from '../opencode/delivery/OpenCodeMemberMessageDeliveryService';
+import type { OpenCodeMemberInboxDelivery } from '../opencode/delivery/OpenCodeMemberMessageDeliveryPorts';
 import type { OpenCodePromptDeliveryLedgerRecord } from '../opencode/delivery/OpenCodePromptDeliveryLedger';
 
 export interface OpenCodeMemberInboxRelayResult {
