+++ src/main/services/team/provisioning/TeamProvisioningOpenCodePromptDeliveryCompatibilityFacade.ts
@@ -1,6 +1,6 @@
 import { createLogger } from '@shared/utils/logger';
 
-import { type OpenCodeRuntimeMessageAdapter } from '../opencode/delivery/OpenCodeMemberMessageDeliveryService';
+import { type OpenCodeRuntimeMessageAdapter } from '../opencode/delivery/OpenCodeMemberMessageDeliveryPorts';
 import {
   type OpenCodePromptDeliveryLedgerRecord,
   type OpenCodePromptDeliveryLedgerStore,
