+++ src/main/services/team/provisioning/TeamProvisioningOpenCodeRuntimeRecoveryFacade.ts
@@ -23,7 +23,7 @@ import {
 import type {
   OpenCodeMemberDirectory,
   OpenCodeMemberIdentityResolution,
-} from '../opencode/delivery/OpenCodeMemberMessageDeliveryService';
+} from '../opencode/delivery/OpenCodeMemberMessageDeliveryPorts';
 import type { TeamProvisioningOpenCodeStoppedLaneCleanupBoundary } from './TeamProvisioningOpenCodeStoppedLaneCleanupBoundary';
 import type { TeamProviderId } from '@shared/types';
