+++ src/main/services/team/provisioning/__tests__/TeamProvisioningOpenCodeRuntimeRecoveryIdentity.test.ts
@@ -9,7 +9,7 @@ import {
 import type {
   OpenCodeMemberDirectory,
   OpenCodeMemberIdentityResolution,
-} from '../../opencode/delivery/OpenCodeMemberMessageDeliveryService';
+} from '../../opencode/delivery/OpenCodeMemberMessageDeliveryPorts';
 import type { OpenCodeRuntimeLaneIndex } from '../../opencode/store/OpenCodeRuntimeManifestEvidenceReader';
 
 function createLaneIndex(lanes: OpenCodeRuntimeLaneIndex['lanes']): OpenCodeRuntimeLaneIndex {
