+++ src/main/services/team/provisioning/__tests__/TeamProvisioningOrgConfigCompatibilityFacade.test.ts
@@ -8,7 +8,7 @@ import {
 import type {
   OpenCodeMemberDirectory,
   OpenCodeMemberIdentityResolution,
-} from '../../opencode/delivery/OpenCodeMemberMessageDeliveryService';
+} from '../../opencode/delivery/OpenCodeMemberMessageDeliveryPorts';
 import type { TeamConfig, TeamMember } from '@shared/types';
 
 describe('TeamProvisioningOrgConfigCompatibilityFacade', () => {
